Implementation of the Last Index of Zero problem.
This is my stab at implementing the Last Index of Zero problem from iJava, where the goal is to take a list of integers containing at least one zero, and return the zero-based index of the last zero in the list.

Most of this was based on a previous problem from thelmuth. The most "creative" part was creating the test suite, which he should definitely look over before pulling this in and using it.