Made my domains vectors, and improved length distribution of random inputs.
I was generating lists instead of vectors for several of my test domains, which apparently breaks the world, and could certainly have contributed to my poor results when I did a few runs on this problem. I also improved the length distribution of randomly generated inputs so there hopefully isn't so much reward for generating a constant that is the roughly half the length of the most dominant input vector length.