Got the basic Midje tests to work.
It turns out by default that midje loads _all_ the files in both `src` and `test`. This dies horribly because many of the files in `src` have side effects on global atoms, so evaluating them all leads to many conflicts. So to avoid this when running midje use `lein midje :autotest test`. This tells midje to only evaluate the files in the `test` sub-directory, ignoring everything in `src`.