Fix releases!
Instead of attaching just a snapshot release, we add a "qualifier".

This means that if we just released `2.7.0`, then we change the
release to be `2.7.0-1-SNAPSHOT` after the release to signal
that we are now working on a release after `2.7.0` that is not
released yet. Then we we do our next release, we bump the version
however much we need (patch, minor, major) and remove the `-1-SNAPSHOT`
to get the version to release on Clojars. Then add it back on after
with the new release.