Undo whitespace changes
This will not make `git blame` more useful, but it will make the diff
more readable when merging into master. The best thing to do would be to
go back and modify commit 015d591836425f3ba6e3256a0e8480c3ae399f07 to
not introduce the whitespace, but I'm not sure that we can do that at
this point.